Fixes frame rate issues and Improved compression.
The frame rate has been changed so that it will always be 30fps when dropping the frame rate. The previous specification was to first drop the frame rate to 1/4, and then to 1/2 if the file size did not reach 8MB or less, but this made no sense and was a waste of time.

Added a message window.
It was hard to tell when the process was finished, so I added it. This is a change for the exe file.

Optimized to work with Pyinstaller.
Pyinstaller was giving an error when using the --no-console option. This was caused by subprocess, and has been fixed.